+++ src/services/gordaApi/WhatsAppClient.ts
@@ -22,7 +22,6 @@ export default class WhatsAppClient implements WPSubject {
     const port = process.env.VUE_APP_WP_CLIENT_API_PORT ?? 3000
 
     this.socket = io( url + ':' + port, {
-      reconnectionAttempts: 5,
       reconnectionDelay: 5000,
       query: {
         clientId: this.wpClient.id
@@ -59,7 +58,7 @@ export default class WhatsAppClient implements WPSubject {
   
   getState(): void {
     this.socket.on(WhatsApp.EVENT_CLIENT, (info) => {
-      if(info) this.socket.emit(WhatsApp.EVENT_GET_STATE)
+      this.socket.emit(WhatsApp.EVENT_GET_STATE)
     })
     this.socket.on(WhatsApp.EVENT_GET_STATE, (state: string) => {
       this.state = state