+++ src/services/stores/MetricsStore.ts
@@ -5,6 +5,9 @@ import DateHelper from '@/helpers/DateHelper'
 import {MetricItem} from '@/types/MetricItem'
 import {ServiceStatus} from '@/types/ServiceStatus'
 import {useLoadingState} from '@/services/stores/LoadingState'
+import MetricRepository from '@/repositories/MetricRepository'
+import { useDriversStore } from './DriversStore'
+import { TopFrequency } from '@/constants/TopFrequency'
 
 const GLOBAL_METRIC_PATH = '/metrics/global'
 
@@ -16,12 +19,33 @@ export const useMetricsStore = defineStore('metricsStore', {
 			completedYearMetric: new Map<string, number>(),
 			canceledYearMetric: new Map<string, number>(),
 			percentYearMetric: new Map<string, number>(),
+			top5DailyMetric: new Map<string, number>(),
 			loaded: false,
+			loading: false
 		}
 	},
 	actions: {
 		getGlobalMetric(startDate: string, endDate: string): Promise<void> {
-			const {setLoading} = useLoadingState()
+			const { setLoading } = useLoadingState()
+			this.loading = true
+			const lastMetricQuery = sessionStorage.getItem('lastMetricQuery')
+			this.globalMetric.splice(0, this.globalMetric.length)
+
+			if (this.metricWasQueryedToday(lastMetricQuery as string)) {
+				const storedMetrics = sessionStorage.getItem('globalMetric')
+				if (storedMetrics) {
+					const parsedMetrics = JSON.parse(storedMetrics) as Metric[]
+					parsedMetrics.forEach((metric) => {
+						this.globalMetric.push(metric)
+					})
+					this.loaded = true
+					return Promise.resolve()
+				} else {
+					this.loaded = false
+					this.loading = true
+				}
+			}
+
 			return new Promise((resolve, reject) => {
 				setLoading(true)
 				axios.get(process.env.VUE_APP_GORDA_API_URL + GLOBAL_METRIC_PATH, {
@@ -30,13 +54,17 @@ export const useMetricsStore = defineStore('metricsStore', {
 						endDate: endDate,
 					},
 				}).then((res) => {
+					this.setMetricQueryToday('lastMetricQuery')
 					setLoading(false)
+					this.loading = false
 					this.loaded = true
 					res.data.data.forEach((metric: Metric) => {
 						this.globalMetric.push(metric)
 					})
+					sessionStorage.setItem('globalMetric', JSON.stringify(this.globalMetric))
 					resolve()
 				}).catch(e => {
+					this.loading = false
 					setLoading(false)
 					console.log(e.message)
 					reject(e)
@@ -46,7 +74,9 @@ export const useMetricsStore = defineStore('metricsStore', {
 		async getCurrentYearMetric(): Promise<void> {
 			const monthLastYear = DateHelper.lastYear()
 			const currentMonth = DateHelper.stringNow()
+			this.loading = true
 			await this.getGlobalMetric(monthLastYear, currentMonth)
+			this.loading = false
 			this.groupAndSumByMonth()
 		},
 		groupAndSumByMonth(): void {
@@ -83,6 +113,54 @@ export const useMetricsStore = defineStore('metricsStore', {
 					metric.set(key, item.amount);
 				}
 			})
+		},
+
+		async getTop5Metric(frequency: TopFrequency): Promise<void> {
+			this.loading = true
+			this.top5DailyMetric.clear()
+			const lastWeeklyMetricQuery = sessionStorage.getItem('lastWeeklyMetricQuery')
+
+			if (frequency === TopFrequency.Weekly && this.metricWasQueryedToday(lastWeeklyMetricQuery as string)) {
+				const storedMetrics = sessionStorage.getItem('top5WeeklyMetric')
+				if (storedMetrics) {
+					const parsedMetrics = JSON.parse(storedMetrics) as [string, number][]
+					parsedMetrics.forEach(([key, value]) => {
+						this.top5DailyMetric.set(key, value)
+					})
+					this.loading = false
+					return Promise.resolve()
+				} else {
+					this.loading = true
+				}
+			}
+
+			const { findById } = useDriversStore()
+			let from = DateHelper.startOfDayUnix()
+			if (frequency === TopFrequency.Weekly) {
+				from = DateHelper.startOfWeekUnix()
+			}
+			const endOfDay = DateHelper.endOfDayUnix()
+			const metrics = await MetricRepository.getDailyTopFive(from, endOfDay)
+			this.loading = false
+				
+			Array.from(metrics.entries()).forEach(([driverId, metric]) => {
+				const driver = findById(driverId)
+				if (driver) this.top5DailyMetric.set(driver.vehicle.plate, metric.count)
+			})
+			
+			if (frequency === TopFrequency.Weekly) {
+				this.setMetricQueryToday('lastWeeklyMetricQuery')
+				sessionStorage.setItem('top5WeeklyMetric', JSON.stringify(Array.from(this.top5DailyMetric.entries())))
+			}
+		},
+
+		setMetricQueryToday(query: string): void {
+			sessionStorage.setItem(query, DateHelper.stringNow())
+		},
+
+		metricWasQueryedToday(query: string): boolean {
+			const today = DateHelper.stringNow()
+			return query === today
 		}
 	},
 })